% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.h.R
\name{arima}
\alias{arima}
\title{Univariate time series}
\usage{
arima(
  data,
  mode = "simple",
  run = FALSE,
  run1 = FALSE,
  dep,
  dep1,
  time1 = NULL,
  freq = 12,
  pred = 10,
  plot = FALSE,
  box = FALSE,
  plot1 = FALSE,
  plot2 = FALSE,
  plot3 = FALSE,
  coef = FALSE,
  fit = FALSE,
  point = FALSE,
  plot4 = FALSE,
  plot5 = FALSE,
  plot6 = FALSE,
  periods = 365,
  unit = "day",
  resid = FALSE,
  plot9 = FALSE,
  plot10 = FALSE,
  level = 95,
  lbLag = 0,
  clean = "FALSE",
  showAcc = FALSE
)
}
\arguments{
\item{data}{timeseries data format}

\item{mode}{.}

\item{run}{.}

\item{run1}{.}

\item{dep}{the dependent variable from \code{data}, variable must be
numeric}

\item{dep1}{the dependent variable from \code{data}, variable must be
numeric}

\item{time1}{ts type called 'ds'}

\item{freq}{.}

\item{pred}{.}

\item{plot}{.}

\item{box}{.}

\item{plot1}{.}

\item{plot2}{.}

\item{plot3}{.}

\item{coef}{.}

\item{fit}{.}

\item{point}{.}

\item{plot4}{.}

\item{plot5}{.}

\item{plot6}{.}

\item{periods}{.}

\item{unit}{.}

\item{resid}{.}

\item{plot9}{.}

\item{plot10}{.}

\item{level}{.}

\item{lbLag}{.}

\item{clean}{.}

\item{showAcc}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$progressBarHTML} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$coef} \tab \tab \tab \tab \tab a table \cr
  \code{results$fit} \tab \tab \tab \tab \tab a table \cr
  \code{results$point} \tab \tab \tab \tab \tab a table \cr
  \code{results$accTrain} \tab \tab \tab \tab \tab a table \cr
  \code{results$accCV} \tab \tab \tab \tab \tab a table \cr
  \code{results$resid$diagTable} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$box} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot3} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot4} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot5} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot6} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot9} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot10} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$coef$asDF}

\code{as.data.frame(results$coef)}
}
\description{
Univariate time series
}
