% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correspondence.h.R
\name{correspondence}
\alias{correspondence}
\title{Correspondence Analysis}
\usage{
correspondence(
  data,
  labels,
  vars,
  nd = 2,
  chi = TRUE,
  eigen = FALSE,
  type = "coordinates",
  type1 = "coordinates",
  col = FALSE,
  row = FALSE,
  plot1 = TRUE,
  plot2 = FALSE,
  plot3 = FALSE,
  plot4 = FALSE
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{labels}{.}

\item{vars}{.}

\item{nd}{.}

\item{chi}{.}

\item{eigen}{.}

\item{type}{.}

\item{type1}{.}

\item{col}{.}

\item{row}{.}

\item{plot1}{.}

\item{plot2}{.}

\item{plot3}{.}

\item{plot4}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$chi} \tab \tab \tab \tab \tab a table \cr
  \code{results$eigen} \tab \tab \tab \tab \tab a table \cr
  \code{results$col} \tab \tab \tab \tab \tab a table \cr
  \code{results$row} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot4} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot3} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$chi$asDF}

\code{as.data.frame(results$chi)}
}
\description{
Correspondence Analysis
}
