% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscan.h.R
\name{dbscan}
\alias{dbscan}
\title{Density-based Clustering}
\usage{
dbscan(
  data,
  vars,
  eps = 0.5,
  minPts = 5,
  standardize = TRUE,
  dist = "euclidean",
  sizeTable = FALSE,
  plot = FALSE,
  profileTable = FALSE,
  prostat = "mean",
  knnPlot = FALSE,
  autoEps = FALSE,
  epsQuantile = 0.95,
  quality = FALSE,
  qoption = "none"
)
}
\arguments{
\item{data}{.}

\item{vars}{.}

\item{eps}{.}

\item{minPts}{.}

\item{standardize}{.}

\item{dist}{.}

\item{sizeTable}{.}

\item{plot}{.}

\item{profileTable}{.}

\item{prostat}{.}

\item{knnPlot}{.}

\item{autoEps}{.}

\item{epsQuantile}{.}

\item{quality}{.}

\item{qoption}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$cm} \tab \tab \tab \tab \tab an output \cr
  \code{results$sizeTable} \tab \tab \tab \tab \tab a table \cr
  \code{results$profileTable} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$knnPlot} \tab \tab \tab \tab \tab an image \cr
  \code{results$autoTable} \tab \tab \tab \tab \tab a table \cr
  \code{results$qualityTable} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$sizeTable$asDF}

\code{as.data.frame(results$sizeTable)}
}
\description{
Density-based Clustering
}
