% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.h.R
\name{group}
\alias{group}
\title{Group plot}
\usage{
group(data, facs, vars, plot = TRUE, plot1 = FALSE)
}
\arguments{
\item{data}{The data as a data frame.}

\item{facs}{.}

\item{vars}{.}

\item{plot}{.}

\item{plot1}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Group plot
}
