% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc.h.R
\name{hc}
\alias{hc}
\title{Clustering Dendrogram}
\usage{
hc(
  mode,
  data,
  labels,
  vars,
  stand = TRUE,
  k = 2,
  metric = "euclidean",
  method = "ward.D2",
  type = "rectangle",
  plot = FALSE,
  horiz = FALSE,
  vars1,
  method1 = "average",
  nb = 100,
  dm = "correlation",
  plot1 = FALSE
)
}
\arguments{
\item{mode}{.}

\item{data}{The data as a data frame.}

\item{labels}{.}

\item{vars}{.}

\item{stand}{.}

\item{k}{.}

\item{metric}{.}

\item{method}{.}

\item{type}{.}

\item{plot}{.}

\item{horiz}{.}

\item{vars1}{.}

\item{method1}{.}

\item{nb}{.}

\item{dm}{.}

\item{plot1}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$progressBarHTML} \tab \tab \tab \tab \tab a html \cr
  \code{results$clust} \tab \tab \tab \tab \tab an output \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$clusterInfo} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$clusterInfo$asDF}

\code{as.data.frame(results$clusterInfo)}
}
\description{
Clustering Dendrogram
}
