% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcm.h.R
\name{hcm}
\alias{hcm}
\title{Hierarchical Clustering}
\usage{
hcm(
  data,
  podatki,
  imena,
  stand = TRUE,
  dis = "euclidean",
  method = "ward.D2",
  grp = "number",
  group = 2,
  hght = 15,
  case = FALSE,
  dend = FALSE,
  horiz = FALSE,
  heat = FALSE,
  pair = FALSE
)
}
\arguments{
\item{data}{.}

\item{podatki}{.}

\item{imena}{.}

\item{stand}{.}

\item{dis}{.}

\item{method}{.}

\item{grp}{.}

\item{group}{.}

\item{hght}{.}

\item{case}{.}

\item{dend}{.}

\item{horiz}{.}

\item{heat}{.}

\item{pair}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$izpis} \tab \tab \tab \tab \tab a table \cr
  \code{results$groups} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$heat} \tab \tab \tab \tab \tab an image \cr
  \code{results$pairs} \tab \tab \tab \tab \tab an image \cr
  \code{results$clust} \tab \tab \tab \tab \tab an output \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$izpis$asDF}

\code{as.data.frame(results$izpis)}
}
\description{
Hierarchical Clustering
}
