% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.h.R
\name{mds}
\alias{mds}
\title{Multidimensional Scaling Plot}
\usage{
mds(
  mode = "simple",
  data,
  labels,
  vars,
  k = 2,
  plot = FALSE,
  plot1 = FALSE,
  stress = FALSE,
  plot3 = FALSE,
  xlab,
  ylab,
  zlab,
  plot2 = FALSE,
  metric = "euclidean"
)
}
\arguments{
\item{mode}{.}

\item{data}{The data as a data frame.}

\item{labels}{.}

\item{vars}{.}

\item{k}{.}

\item{plot}{.}

\item{plot1}{.}

\item{stress}{.}

\item{plot3}{.}

\item{xlab}{.}

\item{ylab}{.}

\item{zlab}{.}

\item{plot2}{.}

\item{metric}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$stressNote} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot3} \tab \tab \tab \tab \tab an image \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$clust} \tab \tab \tab \tab \tab an output \cr
}
}
\description{
Multidimensional Scaling Plot
}
