% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.h.R
\name{pca}
\alias{pca}
\title{Dimension Reduction Plot}
\usage{
pca(
  data,
  mode = "simple",
  labels,
  vars,
  pcaLabels = FALSE,
  eigen = TRUE,
  plot = FALSE,
  plot1 = FALSE,
  plot2 = FALSE,
  vars1,
  facs,
  plot3 = FALSE,
  plot4 = FALSE,
  vars2,
  facs2,
  labels2,
  umapPlot = TRUE,
  umapStandardize = TRUE,
  umapNeighbors = 15,
  umapMinDist = 0.1,
  umapMetric = "euclidean",
  umapSeed = 1234,
  umapLabels = FALSE,
  umapLegend = FALSE
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{mode}{.}

\item{labels}{.}

\item{vars}{.}

\item{pcaLabels}{.}

\item{eigen}{.}

\item{plot}{.}

\item{plot1}{.}

\item{plot2}{.}

\item{vars1}{.}

\item{facs}{.}

\item{plot3}{.}

\item{plot4}{.}

\item{vars2}{.}

\item{facs2}{.}

\item{labels2}{.}

\item{umapPlot}{.}

\item{umapStandardize}{.}

\item{umapNeighbors}{.}

\item{umapMinDist}{.}

\item{umapMetric}{.}

\item{umapSeed}{.}

\item{umapLabels}{.}

\item{umapLegend}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$eigen} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot3} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot4} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot5} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$eigen$asDF}

\code{as.data.frame(results$eigen)}
}
\description{
Dimension Reduction Plot
}
