% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.h.R
\name{roc}
\alias{roc}
\title{ROC Analysis}
\usage{
roc(
  data,
  dep,
  covs,
  plot1 = FALSE,
  plot2 = FALSE,
  plot3 = FALSE,
  auc = FALSE,
  dif = FALSE,
  overall = FALSE,
  positive,
  optimalCutpoint = FALSE,
  specifiedCutpoint = FALSE,
  cutpointValue = 0,
  classificationTable = FALSE
)
}
\arguments{
\item{data}{.}

\item{dep}{.}

\item{covs}{.}

\item{plot1}{.}

\item{plot2}{.}

\item{plot3}{.}

\item{auc}{.}

\item{dif}{.}

\item{overall}{.}

\item{positive}{.}

\item{optimalCutpoint}{.}

\item{specifiedCutpoint}{.}

\item{cutpointValue}{.}

\item{classificationTable}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot3} \tab \tab \tab \tab \tab an image \cr
  \code{results$auc} \tab \tab \tab \tab \tab a table \cr
  \code{results$dif} \tab \tab \tab \tab \tab a table \cr
  \code{results$overall} \tab \tab \tab \tab \tab a table \cr
  \code{results$optimalCutpoint} \tab \tab \tab \tab \tab a table \cr
  \code{results$specifiedCutpoint} \tab \tab \tab \tab \tab a table \cr
  \code{results$classificationTable} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$auc$asDF}

\code{as.data.frame(results$auc)}
}
\description{
ROC Analysis
}
