% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeclust.h.R
\name{timeclust}
\alias{timeclust}
\title{Time Series Clustering}
\usage{
timeclust(
  data,
  k = 2,
  feature,
  item,
  value,
  plot = FALSE,
  plot1 = FALSE,
  summary = FALSE,
  plot3 = FALSE,
  angle = 90
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{k}{.}

\item{feature}{.}

\item{item}{.}

\item{value}{.}

\item{plot}{.}

\item{plot1}{.}

\item{summary}{.}

\item{plot3}{.}

\item{angle}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$clust} \tab \tab \tab \tab \tab an output \cr
  \code{results$clusterTable} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot3} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$clusterTable$asDF}

\code{as.data.frame(results$clusterTable)}
}
\description{
Time Series Clustering
}
