% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.h.R
\name{arima}
\alias{arima}
\title{ARIMA Model}
\usage{
arima(
  data,
  dep,
  covs = NULL,
  time = NULL,
  freq = 12,
  pred = 10,
  plot = TRUE,
  box = FALSE,
  plot1 = FALSE,
  plot2 = FALSE,
  plot3 = FALSE
)
}
\arguments{
\item{data}{timeseries data format}

\item{dep}{the dependent variable from \code{data}, variable must be
numeric}

\item{covs}{the covariates from \code{data}}

\item{time}{the fixed factors from \code{data}}

\item{freq}{.}

\item{pred}{.}

\item{plot}{.}

\item{box}{.}

\item{plot1}{.}

\item{plot2}{.}

\item{plot3}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$box} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot3} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
ARIMA Model
}
