% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.h.R
\name{arima}
\alias{arima}
\title{ARIMA Model}
\usage{
arima(
  data,
  dep,
  covs = NULL,
  time = NULL,
  freq = 12,
  pred = 10,
  plot = TRUE,
  box = FALSE,
  plot1 = FALSE,
  plot2 = FALSE,
  plot3 = FALSE,
  coef = TRUE,
  fit = TRUE,
  point = TRUE
)
}
\arguments{
\item{data}{timeseries data format}

\item{dep}{the dependent variable from \code{data}, variable must be
numeric}

\item{covs}{the covariates from \code{data}}

\item{time}{the fixed factors from \code{data}}

\item{freq}{.}

\item{pred}{.}

\item{plot}{.}

\item{box}{.}

\item{plot1}{.}

\item{plot2}{.}

\item{plot3}{.}

\item{coef}{.}

\item{fit}{.}

\item{point}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$coef} \tab \tab \tab \tab \tab a table \cr
  \code{results$fit} \tab \tab \tab \tab \tab a table \cr
  \code{results$point} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$box} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot3} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$coef$asDF}

\code{as.data.frame(results$coef)}
}
\description{
ARIMA Model
}
