% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.h.R
\name{arima}
\alias{arima}
\title{Univariate time series}
\usage{
arima(
  data,
  mode = "simple",
  dep,
  dep1,
  time1 = NULL,
  freq = 12,
  pred = 10,
  plot = FALSE,
  box = FALSE,
  plot1 = FALSE,
  plot2 = FALSE,
  plot3 = FALSE,
  coef = FALSE,
  fit = FALSE,
  point = FALSE,
  plot4 = FALSE,
  plot5 = FALSE,
  plot6 = FALSE,
  periods = 365,
  unit = "day",
  width1 = 500,
  height1 = 500,
  width2 = 500,
  height2 = 500,
  width3 = 500,
  height3 = 500,
  width4 = 500,
  height4 = 500,
  width5 = 500,
  height5 = 500,
  width6 = 500,
  height6 = 500,
  width7 = 500,
  height7 = 500,
  width8 = 500,
  height8 = 500
)
}
\arguments{
\item{data}{timeseries data format}

\item{mode}{.}

\item{dep}{the dependent variable from \code{data}, variable must be
numeric}

\item{dep1}{the dependent variable from \code{data}, variable must be
numeric}

\item{time1}{ts type called 'ds'}

\item{freq}{.}

\item{pred}{.}

\item{plot}{.}

\item{box}{.}

\item{plot1}{.}

\item{plot2}{.}

\item{plot3}{.}

\item{coef}{.}

\item{fit}{.}

\item{point}{.}

\item{plot4}{.}

\item{plot5}{.}

\item{plot6}{.}

\item{periods}{.}

\item{unit}{.}

\item{width1}{.}

\item{height1}{.}

\item{width2}{.}

\item{height2}{.}

\item{width3}{.}

\item{height3}{.}

\item{width4}{.}

\item{height4}{.}

\item{width5}{.}

\item{height5}{.}

\item{width6}{.}

\item{height6}{.}

\item{width7}{.}

\item{height7}{.}

\item{width8}{.}

\item{height8}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$coef} \tab \tab \tab \tab \tab a table \cr
  \code{results$fit} \tab \tab \tab \tab \tab a table \cr
  \code{results$point} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$box} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot3} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot4} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot5} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot6} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$coef$asDF}

\code{as.data.frame(results$coef)}
}
\description{
Univariate time series
}
