% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret.h.R
\name{caret}
\alias{caret}
\title{Machine Learning}
\usage{
caret(data, dep, group, alt = "notequal", varEq = TRUE)
}
\arguments{
\item{data}{.}

\item{dep}{.}

\item{group}{.}

\item{alt}{.}

\item{varEq}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
}
}
\description{
Machine Learning
}
