% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret.h.R
\name{caret}
\alias{caret}
\title{Machine Learning}
\usage{
caret(
  data,
  dep,
  covs,
  facs,
  trans = "bagImpute",
  method = "pls",
  cm1 = "ctree",
  mecon = "cv",
  number = 10,
  repeats = 5,
  tune = 10,
  per = 0.7,
  over1 = TRUE,
  over = FALSE,
  cla = FALSE,
  cla1 = FALSE,
  tra = TRUE,
  tes = FALSE,
  plot = FALSE,
  plot1 = FALSE,
  plot2 = FALSE,
  plot3 = FALSE,
  plot4 = FALSE,
  plot5 = FALSE,
  plot6 = FALSE,
  plot7 = FALSE,
  ml = "pls,ctree,knn,rf",
  me = "cv",
  num = 10,
  rep = 5,
  tune1 = 10,
  accu = FALSE,
  kapp = FALSE,
  width = 500,
  height = 500,
  width1 = 500,
  height1 = 500,
  width2 = 500,
  height2 = 500,
  width3 = 500,
  height3 = 500,
  width4 = 500,
  height4 = 500,
  width5 = 500,
  height5 = 500,
  width6 = 500,
  height6 = 500,
  width7 = 500,
  height7 = 500,
  plot8 = FALSE,
  width8 = 500,
  height8 = 500
)
}
\arguments{
\item{data}{.}

\item{dep}{.}

\item{covs}{.}

\item{facs}{.}

\item{trans}{.}

\item{method}{.}

\item{cm1}{.}

\item{mecon}{.}

\item{number}{.}

\item{repeats}{.}

\item{tune}{.}

\item{per}{.}

\item{over1}{.}

\item{over}{.}

\item{cla}{.}

\item{cla1}{.}

\item{tra}{.}

\item{tes}{.}

\item{plot}{.}

\item{plot1}{.}

\item{plot2}{.}

\item{plot3}{Binary dependent required.}

\item{plot4}{.}

\item{plot5}{.}

\item{plot6}{.}

\item{plot7}{.}

\item{ml}{.}

\item{me}{.}

\item{num}{.}

\item{rep}{.}

\item{tune1}{.}

\item{accu}{.}

\item{kapp}{.}

\item{width}{.}

\item{height}{.}

\item{width1}{.}

\item{height1}{.}

\item{width2}{.}

\item{height2}{.}

\item{width3}{.}

\item{height3}{.}

\item{width4}{.}

\item{height4}{.}

\item{width5}{.}

\item{height5}{.}

\item{width6}{.}

\item{height6}{.}

\item{width7}{.}

\item{height7}{.}

\item{plot8}{.}

\item{width8}{.}

\item{height8}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text1} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$over1} \tab \tab \tab \tab \tab a table \cr
  \code{results$over} \tab \tab \tab \tab \tab a table \cr
  \code{results$tra} \tab \tab \tab \tab \tab a table \cr
  \code{results$tes} \tab \tab \tab \tab \tab a table \cr
  \code{results$cla1} \tab \tab \tab \tab \tab a table \cr
  \code{results$cla} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot8} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot4} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot3} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot5} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot6} \tab \tab \tab \tab \tab an image \cr
  \code{results$mf$accu} \tab \tab \tab \tab \tab a table \cr
  \code{results$mf$kapp} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot7} \tab \tab \tab \tab \tab an image \cr
  \code{results$pred} \tab \tab \tab \tab \tab an output \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$over1$asDF}

\code{as.data.frame(results$over1)}
}
\description{
Machine Learning
}
