% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret.h.R
\name{caret}
\alias{caret}
\title{Machine Learning}
\usage{
caret(
  data,
  dep,
  covs,
  scale,
  mecon,
  method = "pls",
  cm1 = "knn",
  number = 10,
  repeats = 5,
  tune = 10,
  per = 0.7,
  over1 = TRUE,
  over = FALSE,
  cla = FALSE,
  cla1 = FALSE,
  tra = TRUE,
  tes = FALSE,
  plot = TRUE,
  plot1 = FALSE,
  plot2 = TRUE,
  plot3 = TRUE
)
}
\arguments{
\item{data}{.}

\item{dep}{.}

\item{covs}{.}

\item{scale}{.}

\item{mecon}{.}

\item{method}{.}

\item{cm1}{.}

\item{number}{.}

\item{repeats}{.}

\item{tune}{.}

\item{per}{.}

\item{over1}{.}

\item{over}{.}

\item{cla}{.}

\item{cla1}{.}

\item{tra}{.}

\item{tes}{.}

\item{plot}{.}

\item{plot1}{.}

\item{plot2}{.}

\item{plot3}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$over1} \tab \tab \tab \tab \tab a table \cr
  \code{results$over} \tab \tab \tab \tab \tab a table \cr
  \code{results$tra} \tab \tab \tab \tab \tab a table \cr
  \code{results$tes} \tab \tab \tab \tab \tab a table \cr
  \code{results$cla1} \tab \tab \tab \tab \tab a table \cr
  \code{results$cla} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot3} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$over1$asDF}

\code{as.data.frame(results$over1)}
}
\description{
Machine Learning
}
