% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correspondence.h.R
\name{correspondence}
\alias{correspondence}
\title{CA plot}
\usage{
correspondence(data, labels, vars, plot1 = TRUE, plot2 = FALSE, plot3 = FALSE)
}
\arguments{
\item{data}{The data as a data frame.}

\item{labels}{.}

\item{vars}{.}

\item{plot1}{.}

\item{plot2}{.}

\item{plot3}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot3} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
CA plot
}
