% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correspondence.h.R
\name{correspondence}
\alias{correspondence}
\title{Correspondence Analysis}
\usage{
correspondence(
  data,
  labels,
  vars,
  chi = TRUE,
  eigen = FALSE,
  rowvar = "coordinates",
  loadingind = FALSE,
  colvar = "coordinates",
  loadingvar = FALSE,
  plot1 = TRUE,
  plot2 = FALSE,
  plot3 = FALSE,
  plot4 = FALSE,
  width4 = 500,
  height4 = 500,
  width3 = 500,
  height3 = 500,
  width2 = 500,
  height2 = 500,
  width1 = 500,
  height1 = 500
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{labels}{.}

\item{vars}{.}

\item{chi}{.}

\item{eigen}{.}

\item{rowvar}{.}

\item{loadingind}{.}

\item{colvar}{.}

\item{loadingvar}{.}

\item{plot1}{.}

\item{plot2}{.}

\item{plot3}{.}

\item{plot4}{.}

\item{width4}{.}

\item{height4}{.}

\item{width3}{.}

\item{height3}{.}

\item{width2}{.}

\item{height2}{.}

\item{width1}{.}

\item{height1}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$chi} \tab \tab \tab \tab \tab a table \cr
  \code{results$eigen} \tab \tab \tab \tab \tab a table \cr
  \code{results$loadingvar} \tab \tab \tab \tab \tab a table \cr
  \code{results$loadingind} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot4} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot3} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$chi$asDF}

\code{as.data.frame(results$chi)}
}
\description{
Correspondence Analysis
}
