% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disc.h.R
\name{disc}
\alias{disc}
\title{Linear Discriminant Analysis}
\usage{
disc(
  data,
  dep,
  covs,
  per = 1,
  prior = TRUE,
  gm = FALSE,
  coef = FALSE,
  prop = FALSE,
  tra = FALSE,
  tes = FALSE,
  gc = FALSE,
  plot = FALSE,
  plot1 = FALSE,
  plot2 = FALSE,
  method = "lda",
  width = 500,
  height = 500,
  width1 = 500,
  height1 = 500,
  width2 = 500,
  height2 = 500
)
}
\arguments{
\item{data}{.}

\item{dep}{.}

\item{covs}{.}

\item{per}{.}

\item{prior}{.}

\item{gm}{.}

\item{coef}{.}

\item{prop}{.}

\item{tra}{.}

\item{tes}{.}

\item{gc}{.}

\item{plot}{.}

\item{plot1}{.}

\item{plot2}{.}

\item{method}{.}

\item{width}{.}

\item{height}{.}

\item{width1}{.}

\item{height1}{.}

\item{width2}{.}

\item{height2}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$prior} \tab \tab \tab \tab \tab a table \cr
  \code{results$gm} \tab \tab \tab \tab \tab a table \cr
  \code{results$coef} \tab \tab \tab \tab \tab a table \cr
  \code{results$prop} \tab \tab \tab \tab \tab a table \cr
  \code{results$tra} \tab \tab \tab \tab \tab a table \cr
  \code{results$tes} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$gc} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$prior$asDF}

\code{as.data.frame(results$prior)}
}
\description{
Linear Discriminant Analysis
}
