% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disc.h.R
\name{disc}
\alias{disc}
\title{Linear Discriminant}
\usage{
disc(
  data,
  dep,
  covs,
  homo = FALSE,
  prior = FALSE,
  gm = FALSE,
  coef = FALSE,
  trace = FALSE,
  plot = FALSE
)
}
\arguments{
\item{data}{.}

\item{dep}{.}

\item{covs}{.}

\item{homo}{.}

\item{prior}{.}

\item{gm}{.}

\item{coef}{.}

\item{trace}{.}

\item{plot}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$prior} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$prior$asDF}

\code{as.data.frame(results$prior)}
}
\description{
Linear Discriminant
}
