% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc.h.R
\name{hc}
\alias{hc}
\title{Hierarchical Clustering Dendrogram}
\usage{
hc(data, labels, vars, k = 2, plot = TRUE)
}
\arguments{
\item{data}{The data as a data frame.}

\item{labels}{.}

\item{vars}{.}

\item{k}{.}

\item{plot}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$clust} \tab \tab \tab \tab \tab an output \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Hierarchical Clustering Dendrogram
}
