% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc.h.R
\name{hc}
\alias{hc}
\title{Clustering Dendrogram}
\usage{
hc(
  data,
  labels,
  vars,
  stand = TRUE,
  k = 2,
  metric = "euclidean",
  method = "ward.D2",
  type = "rectangle",
  plot = TRUE
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{labels}{.}

\item{vars}{.}

\item{stand}{.}

\item{k}{.}

\item{metric}{.}

\item{method}{.}

\item{type}{.}

\item{plot}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$clust} \tab \tab \tab \tab \tab an output \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Clustering Dendrogram
}
