% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.h.R
\name{kmeans}
\alias{kmeans}
\title{K-means clustering}
\usage{
kmeans(
  data,
  vars,
  k = 2,
  algo = "Hartigan-Wong",
  nstart = 10,
  plot = TRUE,
  stand = FALSE,
  plot1 = TRUE
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{k}{.}

\item{algo}{.}

\item{nstart}{.}

\item{plot}{.}

\item{stand}{.}

\item{plot1}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$ss} \tab \tab \tab \tab \tab a table \cr
  \code{results$clustering} \tab \tab \tab \tab \tab a table \cr
  \code{results$centroids} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$ss$asDF}

\code{as.data.frame(results$ss)}
}
\description{
K-means clustering
}
