% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.h.R
\name{kmeans}
\alias{kmeans}
\title{K-means Clustering}
\usage{
kmeans(
  data,
  vars,
  factors,
  k = 2,
  k1 = 2,
  max = 10,
  kp = FALSE,
  oc = FALSE,
  algo = "Hartigan-Wong",
  nstart = 10,
  stand = FALSE,
  plot = FALSE,
  angle = 0,
  plot1 = FALSE,
  plot5 = FALSE,
  plot2 = FALSE,
  plot3 = FALSE,
  width = 500,
  height = 500,
  width1 = 500,
  height1 = 500,
  width2 = 500,
  height2 = 500,
  width3 = 500,
  height3 = 500,
  plot4 = FALSE,
  width4 = 500,
  height4 = 500,
  width5 = 500,
  height5 = 500
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{factors}{.}

\item{k}{.}

\item{k1}{.}

\item{max}{.}

\item{kp}{.}

\item{oc}{.}

\item{algo}{.}

\item{nstart}{.}

\item{stand}{.}

\item{plot}{.}

\item{angle}{a number from 0 to 90 defining the angle of the x-axis labels,
where 0 degrees represents completely horizontal labels.}

\item{plot1}{.}

\item{plot5}{.}

\item{plot2}{.}

\item{plot3}{.}

\item{width}{.}

\item{height}{.}

\item{width1}{.}

\item{height1}{.}

\item{width2}{.}

\item{height2}{.}

\item{width3}{.}

\item{height3}{.}

\item{plot4}{.}

\item{width4}{.}

\item{height4}{.}

\item{width5}{.}

\item{height5}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$ss} \tab \tab \tab \tab \tab a table \cr
  \code{results$clustering} \tab \tab \tab \tab \tab a table \cr
  \code{results$centroids} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot5} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot3} \tab \tab \tab \tab \tab an image \cr
  \code{results$clust} \tab \tab \tab \tab \tab an output \cr
  \code{results$clust1} \tab \tab \tab \tab \tab an output \cr
  \code{results$oc} \tab \tab \tab \tab \tab a table \cr
  \code{results$kp} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot4} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$ss$asDF}

\code{as.data.frame(results$ss)}
}
\description{
K-means Clustering
}
