% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long.h.R
\name{long}
\alias{long}
\title{Trajectory K-Means}
\usage{
long(
  data,
  vars,
  run = FALSE,
  nclust = 3,
  standardize = TRUE,
  missing = "listwise",
  nstart = 25,
  itermax = 100,
  algorithm = "Hartigan-Wong",
  seed = 1234,
  showSummary = TRUE,
  showCenters = TRUE,
  showFit = TRUE,
  showPlot = TRUE,
  plotType = "individual",
  lineAlpha = 0.25,
  showElbow = FALSE,
  maxK = 8,
  centerScale = "original",
  showSilhouette = TRUE,
  angle = 0
)
}
\arguments{
\item{data}{.}

\item{vars}{Numeric repeated-measure variables representing time points.}

\item{run}{.}

\item{nclust}{Number of trajectory clusters.}

\item{standardize}{Standardizes repeated-measure variables before
clustering.}

\item{missing}{.}

\item{nstart}{Number of random starting configurations passed to
stats::kmeans().}

\item{itermax}{Maximum number of iterations passed to stats::kmeans().}

\item{algorithm}{K-means algorithm used by stats::kmeans().}

\item{seed}{Random seed for reproducible clustering.}

\item{showSummary}{.}

\item{showCenters}{.}

\item{showFit}{.}

\item{showPlot}{.}

\item{plotType}{.}

\item{lineAlpha}{.}

\item{showElbow}{.}

\item{maxK}{.}

\item{centerScale}{Scale used to display cluster centers.}

\item{showSilhouette}{.}

\item{angle}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$summary} \tab \tab \tab \tab \tab a table \cr
  \code{results$fit} \tab \tab \tab \tab \tab a table \cr
  \code{results$silhouette} \tab \tab \tab \tab \tab a table \cr
  \code{results$centers} \tab \tab \tab \tab \tab a table \cr
  \code{results$member} \tab \tab \tab \tab \tab an output \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$elbow} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$summary$asDF}

\code{as.data.frame(results$summary)}
}
\description{
Trajectory K-Means
}
