% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mca.h.R
\name{mca}
\alias{mca}
\title{MCA Plot}
\usage{
mca(
  data,
  facs,
  vars,
  plot1 = TRUE,
  plot2 = FALSE,
  plot3 = FALSE,
  plot4 = FALSE
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{facs}{.}

\item{vars}{.}

\item{plot1}{.}

\item{plot2}{.}

\item{plot3}{.}

\item{plot4}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot3} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot4} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
MCA Plot
}
