% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mca.h.R
\name{mca}
\alias{mca}
\title{Multiple Correspondence Analysis}
\usage{
mca(
  data,
  facs,
  vars,
  eigen = TRUE,
  plot1 = TRUE,
  plot2 = FALSE,
  plot3 = FALSE,
  plot4 = FALSE
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{facs}{.}

\item{vars}{.}

\item{eigen}{.}

\item{plot1}{.}

\item{plot2}{.}

\item{plot3}{.}

\item{plot4}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$eigen} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot3} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot4} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$eigen$asDF}

\code{as.data.frame(results$eigen)}
}
\description{
Multiple Correspondence Analysis
}
