% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mca.h.R
\name{mca}
\alias{mca}
\title{Multiple Correspondence Analysis}
\usage{
mca(
  data,
  facs,
  vars,
  nd = 2,
  eigen = TRUE,
  rowvar = "coordinates",
  loadingind = FALSE,
  colvar = "coordinates",
  loadingvar = FALSE,
  plot1 = FALSE,
  plot2 = FALSE,
  plot3 = FALSE,
  plot4 = FALSE,
  plot5 = TRUE,
  plot6 = FALSE,
  width1 = 500,
  height1 = 500,
  width2 = 500,
  height2 = 500,
  width3 = 500,
  height3 = 500,
  width4 = 500,
  height4 = 500,
  width5 = 500,
  height5 = 500,
  width6 = 500,
  height6 = 500
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{facs}{.}

\item{vars}{.}

\item{nd}{.}

\item{eigen}{.}

\item{rowvar}{.}

\item{loadingind}{.}

\item{colvar}{.}

\item{loadingvar}{.}

\item{plot1}{.}

\item{plot2}{.}

\item{plot3}{.}

\item{plot4}{.}

\item{plot5}{.}

\item{plot6}{.}

\item{width1}{.}

\item{height1}{.}

\item{width2}{.}

\item{height2}{.}

\item{width3}{.}

\item{height3}{.}

\item{width4}{.}

\item{height4}{.}

\item{width5}{.}

\item{height5}{.}

\item{width6}{.}

\item{height6}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$eigen} \tab \tab \tab \tab \tab a table \cr
  \code{results$loadingvar} \tab \tab \tab \tab \tab a table \cr
  \code{results$loadingind} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot5} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot6} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot3} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot4} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$eigen$asDF}

\code{as.data.frame(results$eigen)}
}
\description{
Multiple Correspondence Analysis
}
