% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.h.R
\name{mds}
\alias{mds}
\title{Multidimensional Scaling Plot}
\usage{
mds(data, labels, vars, plot = FALSE)
}
\arguments{
\item{data}{The data as a data frame.}

\item{labels}{.}

\item{vars}{.}

\item{plot}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Multidimensional Scaling Plot
}
