% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.h.R
\name{mds}
\alias{mds}
\title{Multidimensional Scaling Plot}
\usage{
mds(
  mode = "simple",
  data,
  labels,
  vars,
  k = 2,
  plot = TRUE,
  plot1 = FALSE,
  xlab,
  ylab,
  zlab,
  group,
  plot2 = FALSE
)
}
\arguments{
\item{mode}{.}

\item{data}{The data as a data frame.}

\item{labels}{.}

\item{vars}{.}

\item{k}{.}

\item{plot}{.}

\item{plot1}{.}

\item{xlab}{.}

\item{ylab}{.}

\item{zlab}{.}

\item{group}{.}

\item{plot2}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Multidimensional Scaling Plot
}
