% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.h.R
\name{mfa}
\alias{mfa}
\title{Multiple Factor Analysis}
\usage{
mfa(
  data,
  labels,
  vars,
  k = 2,
  group = "1,3,4,3",
  type = "n,s,s,s",
  gn = "wine_type,Expert1, Expert2,Expert3",
  eigen = TRUE,
  plot = FALSE
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{labels}{.}

\item{vars}{.}

\item{k}{.}

\item{group}{.}

\item{type}{.}

\item{gn}{.}

\item{eigen}{.}

\item{plot}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$eigen} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$eigen$asDF}

\code{as.data.frame(results$eigen)}
}
\description{
Multiple Factor Analysis
}
