% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.h.R
\name{mfa}
\alias{mfa}
\title{Multiple Factor Analysis}
\usage{
mfa(
  data,
  vars,
  facs,
  labels,
  group = "2,5,3,10,9,2",
  type = "n,s,s,s,s,s",
  gn = "a,b,c,d,e,f",
  eigen = TRUE,
  type1 = "coordinates",
  type2 = "coordinates",
  type3 = "coordinates",
  vari = FALSE,
  ind = FALSE,
  quan = FALSE,
  plot = FALSE,
  plot1 = FALSE,
  plot2 = FALSE,
  plot3 = FALSE,
  plot4 = FALSE,
  plot5 = FALSE,
  plot6 = FALSE,
  plot7 = FALSE,
  plot8 = FALSE,
  angle = 0,
  width = 500,
  height = 500,
  width1 = 500,
  height1 = 500,
  width2 = 500,
  height2 = 500,
  width3 = 500,
  height3 = 500,
  width4 = 500,
  height4 = 500,
  width5 = 500,
  height5 = 500,
  width6 = 500,
  height6 = 500,
  width7 = 500,
  height7 = 500,
  width8 = 500,
  height8 = 500
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{facs}{.}

\item{labels}{.}

\item{group}{.}

\item{type}{.}

\item{gn}{.}

\item{eigen}{.}

\item{type1}{.}

\item{type2}{.}

\item{type3}{.}

\item{vari}{.}

\item{ind}{.}

\item{quan}{.}

\item{plot}{.}

\item{plot1}{.}

\item{plot2}{.}

\item{plot3}{.}

\item{plot4}{.}

\item{plot5}{.}

\item{plot6}{.}

\item{plot7}{.}

\item{plot8}{.}

\item{angle}{.}

\item{width}{.}

\item{height}{.}

\item{width1}{.}

\item{height1}{.}

\item{width2}{.}

\item{height2}{.}

\item{width3}{.}

\item{height3}{.}

\item{width4}{.}

\item{height4}{.}

\item{width5}{.}

\item{height5}{.}

\item{width6}{.}

\item{height6}{.}

\item{width7}{.}

\item{height7}{.}

\item{width8}{.}

\item{height8}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$eigen} \tab \tab \tab \tab \tab a table \cr
  \code{results$vari} \tab \tab \tab \tab \tab a table \cr
  \code{results$ind} \tab \tab \tab \tab \tab a table \cr
  \code{results$quan} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot8} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot6} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot7} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot3} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot4} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot5} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$eigen$asDF}

\code{as.data.frame(results$eigen)}
}
\description{
Multiple Factor Analysis
}
