% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.h.R
\name{mfa}
\alias{mfa}
\title{Multiple Factor Analysis}
\usage{
mfa(
  data,
  labels,
  vars,
  group = "2,5,3,10,9,2",
  type = "n,s,s,s,s,s",
  gn = "a,b,c,d,e,f",
  eigen = TRUE,
  cg = FALSE,
  ci = FALSE,
  plot = FALSE,
  plot1 = FALSE,
  plot2 = FALSE,
  plot3 = FALSE,
  plot4 = FALSE,
  plot5 = FALSE,
  plot6 = FALSE,
  plot7 = FALSE
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{labels}{.}

\item{vars}{.}

\item{group}{.}

\item{type}{.}

\item{gn}{.}

\item{eigen}{.}

\item{cg}{.}

\item{ci}{.}

\item{plot}{.}

\item{plot1}{.}

\item{plot2}{.}

\item{plot3}{.}

\item{plot4}{.}

\item{plot5}{.}

\item{plot6}{.}

\item{plot7}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$eigen} \tab \tab \tab \tab \tab a table \cr
  \code{results$cg} \tab \tab \tab \tab \tab a table \cr
  \code{results$ci} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot6} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot7} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot3} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot4} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot5} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$eigen$asDF}

\code{as.data.frame(results$eigen)}
}
\description{
Multiple Factor Analysis
}
