% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.h.R
\name{pca}
\alias{pca}
\title{PCA Plot}
\usage{
pca(data, labels, vars, plot = TRUE, plot1 = FALSE, plot2 = FALSE)
}
\arguments{
\item{data}{The data as a data frame.}

\item{labels}{.}

\item{vars}{.}

\item{plot}{.}

\item{plot1}{.}

\item{plot2}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
PCA Plot
}
