% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.h.R
\name{prophet}
\alias{prophet}
\title{Prophet with Multiple Variables}
\usage{
prophet(data, dep, covs, method = "loess", plot1 = TRUE, plot2 = FALSE)
}
\arguments{
\item{data}{.}

\item{dep}{.}

\item{covs}{.}

\item{method}{.}

\item{plot1}{.}

\item{plot2}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Prophet with Multiple Variables
}
