% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.h.R
\name{roc}
\alias{roc}
\title{ROC Analysis}
\usage{
roc(
  data,
  dep,
  covs,
  plot1 = FALSE,
  width1 = 500,
  height1 = 500,
  plot2 = FALSE,
  width2 = 500,
  height2 = 500,
  plot3 = FALSE,
  width3 = 500,
  height3 = 500
)
}
\arguments{
\item{data}{.}

\item{dep}{.}

\item{covs}{.}

\item{plot1}{.}

\item{width1}{.}

\item{height1}{.}

\item{plot2}{.}

\item{width2}{.}

\item{height2}{.}

\item{plot3}{.}

\item{width3}{.}

\item{height3}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot3} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
ROC Analysis
}
