% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.h.R
\name{roc}
\alias{roc}
\title{ROC Analysis}
\usage{
roc(
  data,
  dep,
  covs,
  plot1 = FALSE,
  width1 = 500,
  height1 = 500,
  plot2 = FALSE,
  width2 = 500,
  height2 = 500,
  plot3 = FALSE,
  width3 = 500,
  height3 = 500,
  auc = FALSE,
  dif = FALSE,
  overall = FALSE
)
}
\arguments{
\item{data}{.}

\item{dep}{.}

\item{covs}{.}

\item{plot1}{.}

\item{width1}{.}

\item{height1}{.}

\item{plot2}{.}

\item{width2}{.}

\item{height2}{.}

\item{plot3}{.}

\item{width3}{.}

\item{height3}{.}

\item{auc}{.}

\item{dif}{.}

\item{overall}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot3} \tab \tab \tab \tab \tab an image \cr
  \code{results$auc} \tab \tab \tab \tab \tab a table \cr
  \code{results$dif} \tab \tab \tab \tab \tab a table \cr
  \code{results$overall} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$auc$asDF}

\code{as.data.frame(results$auc)}
}
\description{
ROC Analysis
}
