% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeclust.h.R
\name{timeclust}
\alias{timeclust}
\title{Time Series Clustering}
\usage{
timeclust(
  data,
  k = 2,
  feature,
  item,
  value,
  plot = FALSE,
  width = 500,
  height = 500
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{k}{.}

\item{feature}{.}

\item{item}{.}

\item{value}{.}

\item{plot}{.}

\item{width}{.}

\item{height}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$clust} \tab \tab \tab \tab \tab an output \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Time Series Clustering
}
