% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.h.R
\name{tree}
\alias{tree}
\title{Decision Tree}
\usage{
tree(
  data,
  dep,
  covs,
  facs,
  per = 0.7,
  plot = FALSE,
  over = FALSE,
  tab = FALSE,
  cla = FALSE,
  over1 = TRUE,
  tab1 = FALSE,
  plot1 = FALSE,
  width = 500,
  height = 500,
  width1 = 500,
  height1 = 500
)
}
\arguments{
\item{data}{.}

\item{dep}{.}

\item{covs}{.}

\item{facs}{.}

\item{per}{.}

\item{plot}{.}

\item{over}{.}

\item{tab}{.}

\item{cla}{.}

\item{over1}{.}

\item{tab1}{.}

\item{plot1}{.}

\item{width}{.}

\item{height}{.}

\item{width1}{.}

\item{height1}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$over1} \tab \tab \tab \tab \tab a table \cr
  \code{results$tab1} \tab \tab \tab \tab \tab a table \cr
  \code{results$over} \tab \tab \tab \tab \tab a table \cr
  \code{results$tab} \tab \tab \tab \tab \tab a table \cr
  \code{results$cla} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$over1$asDF}

\code{as.data.frame(results$over1)}
}
\description{
Decision Tree
}
