% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.h.R
\name{tree}
\alias{tree}
\title{Classification Analysis}
\usage{
tree(data, dep, covs, tw = 1, plot = FALSE)
}
\arguments{
\item{data}{.}

\item{dep}{.}

\item{covs}{.}

\item{tw}{.}

\item{plot}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Classification Analysis
}
